♻️ refactor(move): shared resources now in fetures
