+++ frontend/features/auth/components/LoginForm.tsx
@@ -14,7 +14,7 @@ import { PasswordlessLoginForm } from '@/features/auth/components/PasswordlessFo
 import { useFormState } from 'react-dom'
 import { useSignInState } from '@/features/auth/hooks/use-signIn-state'
 import { cn } from '@/utils/cn'
-import { SubmitButton } from '@/components/shared/SubmitButton'
+import { SubmitButton } from '@/features/shared/SubmitButton'
 import { ForgotPasswordModal } from './ForgotPasswordModal'
 import Link from 'next/link'
