+++ frontend/features/profile/components/SecuritySetting.tsx
@@ -6,7 +6,7 @@ import React from 'react'
 import { Card, CardHeader, CardBody, Button } from '@nextui-org/react'
 import { Icon } from '@iconify/react'
 
-import { SwitchCell } from '@/components/shared/SwitchCell'
+import { SwitchCell } from '@/features/shared/SwitchCell'
 import { cn } from '@/utils/cn'
 import { useUser } from '@/features/user/hooks/use-user'
 import { DeleteAccountModal } from '@/features/profile/components/modal/DeleteAccountModal'
